+++ panel/src/App.tsx
@@ -2,6 +2,7 @@ import { AbortOperationButton } from '@/component/AbortOperationButton'
 import { BranchSelector } from '@/component/BranchSelector'
 import { ComparePanel } from '@/component/ComparePanel'
 import { DragOverlay } from '@/component/DragOverlay'
+import { FloatingScrollActions } from '@/component/FloatingScrollActions'
 import { GitActionLoadingToast } from '@/component/GitActionLoadingToast'
 import { Graph } from '@/component/Graph'
 import { ReflogPanel } from '@/component/ReflogPanel'
@@ -12,6 +13,7 @@ import { UndoButton } from '@/component/UndoButton'
 import { WorktreeMenu } from '@/component/WorktreeMenu'
 import { CompareProvider } from '@/context/CompareContext'
 import { DragProvider } from '@/context/DragContext'
+import { GraphScrollProvider } from '@/context/GraphScrollContext'
 import { SearchProvider } from '@/context/SearchContext'
 import { SettingsProvider } from '@/context/SettingsContext'
 import { ToastProvider } from '@/context/ToastContext'
@@ -109,6 +111,8 @@ const RepoPanel: FC = () => {
           <Graph selectedBranches={selectedBranches} searchTerm={searchTerm} scrollRef={scrollRef} />
         </main>
 
+        <FloatingScrollActions scrollRef={scrollRef} />
+
         <DragOverlay />
 
         <ComparePanel />
@@ -167,11 +171,13 @@ export const App: FC = () => {
   return (
     <QueryClientProvider client={queryClient}>
       <SettingsProvider>
-        <ToastProvider>
-          <DragProvider>
-            <Panel />
-          </DragProvider>
-        </ToastProvider>
+        <GraphScrollProvider>
+          <ToastProvider>
+            <DragProvider>
+              <Panel />
+            </DragProvider>
+          </ToastProvider>
+        </GraphScrollProvider>
       </SettingsProvider>
     </QueryClientProvider>
   )